+++ test/tasks.js
@@ -1,0 +1,408 @@
+import _ from 'underscore';
+import { expect } from 'chai';
+import ApifyClient from '../build';
+import { BASE_PATH } from '../build/tasks';
+import { mockRequest, requestExpectCall, requestExpectErrorCall, restoreRequest } from './_helper';
+
+const BASE_URL = 'http://example.com/something';
+const OPTIONS = { baseUrl: BASE_URL };
+
+describe('Tasks method', () => {
+    before(mockRequest);
+    after(restoreRequest);
+
+    it('listTasks() works', () => {
+        const callOptions = {
+            token: 'sometoken',
+            limit: 5,
+            offset: 3,
+            desc: true,
+        };
+
+        const queryString = {
+            token: 'sometoken',
+            limit: 5,
+            offset: 3,
+            desc: 1,
+        };
+
+        const expected = {
+            limit: 5,
+            offset: 3,
+            count: 5,
+            total: 10,
+            desc: true,
+            items: ['task1', 'task2'],
+        };
+
+        requestExpectCall({
+            json: true,
+            method: 'GET',
+            url: `${BASE_URL}${BASE_PATH}`,
+            qs: queryString,
+        }, {
+            data: expected,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .listTasks(callOptions)
+            .then(response => expect(response).to.be.eql(expected));
+    });
+
+    it('listTasks() works without pagination', () => {
+        const callOptions = {
+            token: 'sometoken',
+        };
+
+        const queryString = {
+            token: 'sometoken',
+        };
+
+        const expected = {
+            limit: 1000,
+            offset: 0,
+            count: 2,
+            total: 2,
+            desc: false,
+            items: ['task1', 'task2'],
+        };
+
+        requestExpectCall({
+            json: true,
+            method: 'GET',
+            url: `${BASE_URL}${BASE_PATH}`,
+            qs: queryString,
+        }, {
+            data: expected,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .listTasks(callOptions)
+            .then(response => expect(response).to.be.eql(expected));
+    });
+
+    it('createTask() works', () => {
+        const task = { foo: 'bar' };
+        const token = 'some-token';
+
+        requestExpectCall({
+            json: true,
+            method: 'POST',
+            url: `${BASE_URL}${BASE_PATH}`,
+            qs: { token },
+            body: task,
+        }, {
+            data: task,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .createTask({ task, token })
+            .then(response => expect(response).to.be.eql(task));
+    });
+
+    it('updateTask() works with both taskId parameter and taskId in task object', () => {
+        const taskId = 'some-user/some-id';
+        const task = { id: taskId, foo: 'bar' };
+        const token = 'some-token';
+
+        requestExpectCall({
+            json: true,
+            method: 'PUT',
+            url: `${BASE_URL}${BASE_PATH}/some-user~some-id`,
+            qs: { token },
+            body: _.omit(task, 'id'),
+        }, {
+            data: task,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .updateTask({ taskId, task, token })
+            .then(response => expect(response).to.be.eql(task));
+    });
+
+    it('updateTask() works with taskId in task object', () => {
+        const taskId = 'some-id';
+        const task = { id: taskId, foo: 'bar' };
+        const token = 'some-token';
+
+        requestExpectCall({
+            json: true,
+            method: 'PUT',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}`,
+            qs: { token },
+            body: _.omit(task, 'id'),
+        }, {
+            data: task,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .updateTask({ task, token })
+            .then(response => expect(response).to.be.eql(task));
+    });
+
+    it('updateTask() works with taskId parameter', () => {
+        const taskId = 'some-id';
+        const task = { foo: 'bar' };
+        const token = 'some-token';
+
+        requestExpectCall({
+            json: true,
+            method: 'PUT',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}`,
+            qs: { token },
+            body: task,
+        }, {
+            data: task,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .updateTask({ taskId, task, token })
+            .then(response => expect(response).to.be.eql(task));
+    });
+
+    it('updateTask() works with taskId as part task.id parameter', () => {
+        const task = { id: 'some-id', foo: 'bar' };
+        const token = 'some-token';
+
+        requestExpectCall({
+            json: true,
+            method: 'PUT',
+            url: `${BASE_URL}${BASE_PATH}/${task.id}`,
+            qs: { token },
+            body: { foo: 'bar' },
+        }, {
+            data: task,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .updateTask({ task, token })
+            .then(response => expect(response).to.be.eql(task));
+    });
+
+    it('getTask() works', () => {
+        const taskId = 'some-id';
+        const task = { id: taskId, foo: 'bar' };
+        const token = 'some-token';
+
+        requestExpectCall({
+            json: true,
+            method: 'GET',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}`,
+            qs: { token },
+        }, {
+            data: task,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .getTask({ taskId, token })
+            .then(response => expect(response).to.be.eql(task));
+    });
+
+    /*
+    it('getTask() works without token', () => {
+        const taskId = 'some-id';
+        const task = { id: taskId, foo: 'bar' };
+
+        requestExpectCall({
+            json: true,
+            method: 'GET',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}`,
+            qs: {},
+        }, {
+            data: task,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .getTask({ taskId })
+            .then(response => expect(response).to.be.eql(task));
+    });
+    */
+
+    it('getTask() returns null on 404 status code (RECORD_NOT_FOUND)', () => {
+        const taskId = 'some-id';
+        const token = 'some-token';
+
+        requestExpectErrorCall({
+            json: true,
+            method: 'GET',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}`,
+            qs: { token },
+        }, false, 404);
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .getTask({ taskId, token })
+            .then(given => expect(given).to.be.eql(null));
+    });
+
+    it('deleteTask() works', () => {
+        const taskId = 'some-id';
+        const token = 'some-token';
+
+        requestExpectCall({
+            json: true,
+            method: 'DELETE',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}`,
+            qs: { token },
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .deleteTask({ taskId, token });
+    });
+
+    it('listRuns() works', () => {
+        const taskId = 'some-id';
+
+        const callOptions = {
+            token: 'sometoken',
+            limit: 5,
+            offset: 3,
+            desc: true,
+        };
+
+        const queryString = {
+            token: 'sometoken',
+            limit: 5,
+            offset: 3,
+            desc: 1,
+        };
+
+        const expected = {
+            limit: 5,
+            offset: 3,
+            desc: true,
+            count: 5,
+            total: 10,
+            items: ['run1', 'run2'],
+        };
+
+        requestExpectCall({
+            json: true,
+            method: 'GET',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}/runs`,
+            qs: queryString,
+        }, {
+            data: expected,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .listRuns(Object.assign({}, callOptions, { taskId }))
+            .then(response => expect(response).to.be.eql(expected));
+    });
+
+    it('listRuns() works without pagination params', () => {
+        const taskId = 'some-id';
+
+        const callOptions = {
+            token: 'sometoken',
+        };
+
+        const queryString = {
+            token: 'sometoken',
+        };
+
+        const expected = {
+            limit: 1000,
+            offset: 0,
+            desc: false,
+            count: 2,
+            total: 2,
+            items: ['run1', 'run2'],
+        };
+
+        requestExpectCall({
+            json: true,
+            method: 'GET',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}/runs`,
+            qs: queryString,
+        }, {
+            data: expected,
+        });
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .listRuns(Object.assign({}, callOptions, { taskId }))
+            .then(response => expect(response).to.be.eql(expected));
+    });
+
+    it('runTask() works', () => {
+        const taskId = 'some-id';
+        const token = 'some-token';
+        const run = { foo: 'bar' };
+        const apiResponse = JSON.stringify({ data: run });
+
+        const waitForFinish = 120;
+
+        requestExpectCall({
+            method: 'POST',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}/runs`,
+            qs: { token, waitForFinish },
+        }, apiResponse);
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .runTask({ taskId, token, waitForFinish })
+            .then(response => expect(response).to.be.eql(run));
+    });
+
+    /*
+
+    it('runTask() works without token', () => {
+        const taskId = 'some-id';
+        const run = { foo: 'bar' };
+        const apiResponse = JSON.stringify({ data: run });
+
+        requestExpectCall({
+            method: 'POST',
+            url: `${BASE_URL}${BASE_PATH}/${taskId}/runs`,
+            qs: {},
+        }, apiResponse);
+
+        const apifyClient = new ApifyClient(OPTIONS);
+
+        return apifyClient
+            .tasks
+            .runTask({ taskId })
+            .then(response => expect(response).to.be.eql(run));
+    });
+    */
+});